refactor : 폴더 기능별 구조 분리
