+++ src/features/loginPage/components/common/AuthFooter.tsx
