+++ src/features/loginPage/components/common/AuthSideCard.tsx
