+++ src/features/loginPage/components/verification/CaptchaBox.tsx
